허수아비 튜토리얼을 수정한다
허수아비 튜토리얼에서 '허수아비'는 다른 튜토리얼에서도 사용하는 객체이기 때문에,
허수아비를 따로 클래스로 정의한다!
