+++ Assets/Scripts/Tutorial/HitBoxObject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a2c399462b90c5840ac678021988731c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
